Avoid the 2GB limit on file sizes
This converts all the internals of KSPTL to use a custom
LargeNativeArray type which can contain more than 2^32-1 elements.
This should allow loading 16k cubemaps, and other textures which might
be larger than supported by unity.

+++ src/KSPTextureLoader/Format/PNGLoader.cs
@@ -63,7 +63,7 @@ internal static class PNGLoader
         var readInfo = new FileLoader.FileReadInfo
         {
             path = diskPath,
-            length = (int)length,
+            length = length,
             offset = 0,
         };
         var task = FileLoader
